handled mockFiles in FileTable component
